backend/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0

# Database
asyncpg>=0.30.0
sqlalchemy[asyncio]>=2.0.36
alembic>=1.14.0
greenlet>=3.1.0

# HTTP client
httpx>=0.28.0
tenacity>=9.0.0

# Background tasks
apscheduler>=3.10.0

# Configuration
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.9.0
